requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0

certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-better-admin-arrayfield==1.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
psycopg2==2.8.5
PyJWT==1.7.1
python-http-client==3.2.7
pytz==2020.1
requests==2.24.0
sendgrid==6.4.1
six==1.15.0
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
twilio==6.43.0
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
